% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\docType{data}
\name{batting}
\alias{batting}
\title{Individual Batting Statistics}
\format{Batting.csv file from the Lahman database}
\usage{
batting
}
\description{
Player batting statistics by season and team from 1871 to 2015
}
\keyword{baseball}
\keyword{batting}
\keyword{lahman}

