% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exRawSthdAdultData.R
\docType{data}
\name{master}
\alias{master}
\title{A registry of all baseball players and managers}
\format{The Master.csv data from the Lahman database}
\usage{
master
}
\description{
This is the Master.csv file from the Lahman database originally saved as a .csv file
}
\keyword{lahman}
\keyword{master}
\keyword{baseball}

