% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\docType{data}
\name{pitching}
\alias{pitching}
\title{Individual Pitching Statistics}
\format{Pitching.csv file from the Lahman database}
\usage{
pitching
}
\description{
Season-by-season pitching data for players
}
\keyword{baseball}
\keyword{lahman}
\keyword{pitching}

