% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\docType{data}
\name{master}
\alias{master}
\title{A Registry of Baseball People}
\format{Master.csv file from the Lahman database}
\usage{
master
}
\description{
The Master table is a registry of every player and manager who have appeared in the MLB. From Lahman database.
}
\keyword{baseball}
\keyword{lahman}
\keyword{master}

